feat: add experimentation logic
