--- homepage/scripts/experimentation-loader.js
@@ -1,100 +1,0 @@
-import { getLibs } from './utils.js';
-
-/**
- * Loads a CSS file.
- * @param {string} href URL to the CSS file
- */
-async function loadCSS(href) {
-  return new Promise((resolve, reject) => {
-    if (!document.querySelector(`head > link[href="${href}"]`)) {
-      const link = document.createElement('link');
-      link.rel = 'stylesheet';
-      link.href = href;
-      link.onload = resolve;
-      link.onerror = reject;
-      document.head.append(link);
-    } else {
-      resolve();
-    }
-  });
-}
-
-/**
- * Gets all the metadata elements that are in the given scope.
- * @param {String} scope The scope/prefix for the metadata
- * @param {Document} doc Document object to query for metadata. Defaults to the window's document
- * @returns an array of HTMLElement nodes that match the given scope
- */
-function getAllMetadata(scope, doc = document) {
-  return [...doc.head.querySelectorAll(`meta[property^="${scope}:"],meta[name^="${scope}-"]`)]
-    .reduce((res, meta) => {
-      const id = toClassName(meta.name
-        ? meta.name.substring(scope.length + 1)
-        : meta.getAttribute('property').split(':')[1]);
-      res[id] = meta.getAttribute('content');
-      return res;
-    }, {});
-}
-
-/**
- * Sanitizes a string for use as class name.
- * @param {string} name The unsanitized string
- * @returns {string} The class name
- */
-function toClassName(name) {
-  return typeof name === 'string'
-    ? name
-      .toLowerCase()
-      .replace(/[^0-9a-z]/gi, '-')
-      .replace(/-+/g, '-')
-      .replace(/^-|-$/g, '')
-    : '';
-}
-
-/**
- * Sanitizes a string for use as a js property name.
- * @param {string} name The unsanitized string
- * @returns {string} The camelCased name
- */
-function toCamelCase(name) {
-  return toClassName(name).replace(/-([a-z])/g, (g) => g[1].toUpperCase());
-}
-
-
-const pluginContext = {
-  getAllMetadata,
-  loadCSS,
-  toCamelCase,
-  toClassName,
-};
-
-export async function setExperimentsContext(codeBasePath) {
-  const miloLibs = getLibs();
-  const { getMetadata } = await import(`${miloLibs}/utils/utils.js`);
-  const { sampleRUM } = await import(`${miloLibs}/utils/samplerum.js`);
-  pluginContext.getMetadata = getMetadata;
-  pluginContext.sampleRUM = sampleRUM;
-  window.hlx.codeBasePath = codeBasePath;
-  window.hlx.patchBlockConfig = [];
-}
-
-export async function runExperiments(config) {
-  if (pluginContext.getMetadata('experiment')
-    || Object.keys(getAllMetadata('campaign')).length
-    || Object.keys(getAllMetadata('audience')).length) {
-    // eslint-disable-next-line import/no-relative-packages
-    const { loadEager: runEager } = await import('../plugins/experimentation/src/index.js');
-    return runEager(document, config, pluginContext);
-  }
-}
-
-export async function showExperimentsOverlay(config) {
-  const miloLibs = getLibs();
-  if (pluginContext.getMetadata('experiment')
-    || Object.keys(getAllMetadata('campaign')).length
-    || Object.keys(getAllMetadata('audience')).length) {
-    // eslint-disable-next-line import/no-relative-packages
-    const { loadLazy: runLazy } = await import('../plugins/experimentation/src/index.js');
-    return runLazy(document, config, pluginContext);
-  }
-}
